+++ src/routes/__tests__/ui.test.js
@@ -221,6 +221,26 @@ describe('Test ui responses', () => {
       })
     })
 
+    describe('Test /provinces/QB response', () => {
+      test('it should return 301', async () => {
+        const response = await request(app).get('/provinces/QB')
+        expect(response.statusCode).toBe(301)
+        expect(response.headers.location).toEqual('/provinces/QC')
+      })
+
+      test('it should return 301 for lowercased provinceId', async () => {
+        const response = await request(app).get('/provinces/qb')
+        expect(response.statusCode).toBe(301)
+        expect(response.headers.location).toEqual('/provinces/QC')
+      })
+
+      test('it should return 301 with the year included', async () => {
+        const response = await request(app).get('/provinces/QB/2022')
+        expect(response.statusCode).toBe(301)
+        expect(response.headers.location).toEqual('/provinces/QC/2022')
+      })
+    })
+
     describe('Test /provinces/:provinceId/:year responses', () => {
       const nextYearMB = getCurrentHolidayYear('MB') + 1